requirements.txt
flask
flask-sqlalchemy
flask-wtf
gunicorn